style: change language switch
