+++ src/components/ui/locale-switch.tsx
@@ -1,0 +1,28 @@
+import * as SwitchPrimitives from "@radix-ui/react-switch";
+import * as React from "react";
+
+import { cn } from "@/lib/utils";
+
+const Switch = React.forwardRef<
+  React.ElementRef<typeof SwitchPrimitives.Root>,
+  React.ComponentPropsWithoutRef<typeof SwitchPrimitives.Root>
+>(({ className, ...props }, ref) => (
+  <SwitchPrimitives.Root
+    className={cn(
+      "peer inline-flex h-7 w-14 shrink-0 cursor-pointer items-center rounded-full border-1 border-secondary shadow-sm transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background disabled:cursor-not-allowed disabled:opacity-50 bg-secondary/50 data-[state=checked]:before:content-['VI'] data-[state=unchecked]:before:content-['EN'] relative before:absolute data-[state=checked]:before:left-2 data-[state=unchecked]:before:right-2 before:text-xs",
+      className
+    )}
+    {...props}
+    ref={ref}
+  >
+    <SwitchPrimitives.Thumb
+      className={cn(
+        "pointer-events-none block h-6 w-6 data-[state=checked]:bg-[url('/assets/vi-flag.svg')] data-[state=unchecked]:bg-[url('/assets/uk-flag.svg')] rounded-full bg-background shadow-lg ring-0 transition-transform data-[state=checked]:translate-x-7 data-[state=unchecked]:translate-x-0"
+      )}
+      style={{ backgroundSize: "cover" }}
+    />
+  </SwitchPrimitives.Root>
+));
+Switch.displayName = SwitchPrimitives.Root.displayName;
+
+export { Switch };
